[COTLMPSERVER] Move the Utils class to main directory of the server
Since the util methods can be used outside of the messages scope, just move it to the main directory of the server.
